move CompositeReceipt to composite.rs
